services/archiver/requirements.txt
cloudevents==1.2.0
opencensus==0.7.13
opencensus-ext-stackdriver==0.8.0
Flask==2.0.1
gunicorn==20.1.0
google-cloud-storage==1.42.2